fix: [mac][ios] address compact density review
